refactor: develop과 동기화 및 ts, vite로 환경 이전
