+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <!-- 폰트 추가가 -->
+    <!-- 폰트 추가 -->
     <link
       href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@400;900&display=swap"
       rel="stylesheet"
