--- my-app/src/pages/Home/index.js
@@ -1,9 +1,0 @@
-import Animation3D from './Animation3D';
-
-export default function Home() {
-  return (
-    <>
-      <Animation3D />
-    </>
-  );
-}
